+++ test/sutra_ui/dialog_test.exs
@@ -7,7 +7,7 @@ defmodule SutraUI.DialogTest do
   alias SutraUI.Dialog
 
   describe "dialog/1 rendering" do
-    test "renders as native dialog element" do
+    test "renders as div-based dialog (screen share compatible)" do
       assigns = %{}
 
       html =
@@ -18,8 +18,9 @@ defmodule SutraUI.DialogTest do
         </Dialog.dialog>
         """)
 
-      assert html =~ "<dialog"
-      assert html =~ "dialog"
+      assert html =~ ~s(<div)
+      assert html =~ ~s(class="dialog)
+      assert html =~ ~s(role="dialog")
     end
 
     test "renders with correct id" do
@@ -106,8 +107,39 @@ defmodule SutraUI.DialogTest do
     end
   end
 
+  describe "dialog/1 show attribute" do
+    test "adds is-open class when show is true" do
+      assigns = %{}
+
+      html =
+        rendered_to_string(~H"""
+        <Dialog.dialog id="test-dialog" show={true}>
+          <:title>Title</:title>
+          Content
+        </Dialog.dialog>
+        """)
+
+      assert html =~ ~s(class="dialog is-open")
+    end
+
+    test "does not add is-open class when show is false" do
+      assigns = %{}
+
+      html =
+        rendered_to_string(~H"""
+        <Dialog.dialog id="test-dialog" show={false}>
+          <:title>Title</:title>
+          Content
+        </Dialog.dialog>
+        """)
+
+      assert html =~ ~s(class="dialog)
+      refute html =~ "is-open"
+    end
+  end
+
   describe "dialog/1 custom class" do
-    test "includes custom class" do
+    test "includes custom class on panel" do
       assigns = %{}
 
       html =
@@ -152,7 +184,22 @@ defmodule SutraUI.DialogTest do
       assert html =~ ~s(aria-describedby="test-dialog-description")
     end
 
-    test "renders close button with aria-label" do
+    test "has role=dialog and aria-modal=true" do
+      assigns = %{}
+
+      html =
+        rendered_to_string(~H"""
+        <Dialog.dialog id="test-dialog">
+          <:title>Title</:title>
+          Content
+        </Dialog.dialog>
+        """)
+
+      assert html =~ ~s(role="dialog")
+      assert html =~ ~s(aria-modal="true")
+    end
+
+    test "renders close button with aria-label when on_cancel is set" do
       assigns = %{}
 
       html =
@@ -164,23 +211,41 @@ defmodule SutraUI.DialogTest do
         """)
 
       assert html =~ ~s(aria-label="Close")
-      assert html =~ ~s(type="submit")
+      assert html =~ ~s(type="button")
     end
   end
 
-  describe "dialog/1 cancel event" do
-    test "renders with on_cancel attribute" do
+  describe "dialog/1 backdrop" do
+    test "renders backdrop element" do
       assigns = %{}
 
       html =
         rendered_to_string(~H"""
-        <Dialog.dialog id="test-dialog" on_cancel="close_dialog">
+        <Dialog.dialog id="test-dialog">
           <:title>Title</:title>
           Content
         </Dialog.dialog>
         """)
 
-      assert html =~ ~s(on_cancel="close_dialog")
+      assert html =~ ~s(class="dialog-backdrop")
+      assert html =~ ~s(aria-hidden="true")
+    end
+  end
+
+  describe "dialog/1 phx-hook" do
+    test "has phx-hook for dialog behavior" do
+      assigns = %{}
+
+      html =
+        rendered_to_string(~H"""
+        <Dialog.dialog id="test-dialog" on_cancel="close">
+          <:title>Title</:title>
+          Content
+        </Dialog.dialog>
+        """)
+
+      assert html =~ "phx-hook="
+      assert html =~ "Dialog"
     end
   end
 
@@ -203,22 +268,6 @@ defmodule SutraUI.DialogTest do
     test "hide_dialog with existing JS struct" do
       js = Phoenix.LiveView.JS.push("some-event") |> Dialog.hide_dialog("test-dialog")
       assert %Phoenix.LiveView.JS{} = js
-    end
-  end
-
-  describe "dialog/1 phx-hook" do
-    test "has phx-hook for dialog behavior" do
-      assigns = %{}
-
-      html =
-        rendered_to_string(~H"""
-        <Dialog.dialog id="test-dialog" on_cancel="close">
-          <:title>Title</:title>
-          Content
-        </Dialog.dialog>
-        """)
-
-      assert html =~ "phx-hook"
     end
   end
 end